pyproject.toml
[tool.poetry]
name = "spotify_py_sdk"
version = "0.1.0"
description = "Python SDK for Spotify"
authors = ["Ananya Nayak <ananyanayak102@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"


[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.0"
pytest = "^7.4.3"
sphinx = "^7.2.6"
sphinx-rtd-theme = "^2.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
